Moving start Appium server method to DriverFactory
